Store.init takes optional middlewares param
dispatcher is generated with the original dispatch func (with a little
twist, the `store` param) and also the reversed middlewares.

We also don't need hold onto the `reducer` anymore. It's wrapped and
referred by the dispatcher closure.

With the current method to generate dispatcher, imagine we have three
middlewares, A, B and C, as long as all middleware calls the passed in
dispatcher, we will have the dispatcher in the store like this:
{ // CALL A
    { // CALL B
        { // CALL C
            { // CALL the original dispatch func
                store.state = reducer(store.state, action)
                store.subscribers.forEach { $0(store) }
            }
            // C: get new state
        }
        // B: get new state
    }
    // A: get new state
}